pyproject.toml
[project]
name = "ea-workbench"
version = "0.1.0"
description = "Stanmore PFC — Enterprise Architect operational workbench"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.0",
    "anthropic>=0.40",
    "claude-code-sdk>=0.0.10",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "lxml>=5.0",
    "aiofiles>=24.0",
    "sqlalchemy>=2.0",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pyright>=1.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ea_workbench"]
